feat: enhance OpenAPI schema generation with SwaggerConfig and OAuth2 support
